+++ app/models/comment.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, BigInteger, ForeignKey, Text
+from sqlalchemy import Column, Integer, ForeignKey, Text
 from sqlalchemy.orm import relationship
 from app.database import Base
 
 class Comment(Base):
     __tablename__ = "comments"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
     post_id = Column(Integer, ForeignKey("posts.id", ondelete="CASCADE"), nullable=False)
     comment = Column(Text, nullable=False)
